swisper/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
PyYAML
jsonschema
python-multipart
shelve
langdetect
emoji
farm-haystack[all]
pytest
pytest-asyncio
